src/requirements.txt
alabaster==0.7.12
asteval==0.9.17
astropy==4.3.post1
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
cached-property==1.5.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
colorspacious==1.1.2
commonmark==0.9.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
h5py==3.3.0
idna==2.8
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==0.23
imreg-dft==2.0.0
intel-openmp==2019.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
kiwisolver==1.1.0
llvmlite==0.36.0
lmfit==0.9.14
magpylib==2.3.0b0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl==2019.0
more-itertools==7.2.0
mpld3==0.3
mpldatacursor==0.6.2
mpmath==1.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numba==0.53.1
numpy==1.21.1
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
packaging==19.2
pandas==1.3.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
plasmapy==0.6.0
plotly==4.4.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
purepng==0.2.0
pyerfa==2.0.0
Pygments==2.4.2
pyparsing==2.4.2
pypng==0.0.20
pypylon==1.4.0
pypylon-opencv-viewer==1.1.0
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
pyusb==1.0.0a2
PyVISA==1.11.3
PyWavelets==1.1.1
pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
recommonmark==0.6.0
requests==2.22.0
retrying==1.3.3
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.3.1
scikit-image==0.16.2
scipy==1.3.3
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.62.0
traitlets==4.3.3
typing-extensions==3.7.4.3
uncertainties==3.1.2
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xarray==0.19.0
xlwt==1.3.0
yolk3k==0.9
zipp==0.6.0
